orders/src/routes/show.ts
import express, { request, Request, Response } from 'express';
import {
  NotAuthorizedError,
  NotFoundError,
  requireAuth,
} from '@mpqticket/common';
import { Order } from '../models/order';
const router = express.Router();

router.get(
  '/api/orders/:id',
  requireAuth,
  async (req: Request, res: Response) => {
    // const order = await Order.findOne({
    //   _id: req.params.id,
    //   userId: req.currentUser!.id,
    // }).populate('ticket');

    const order = await Order.findById(req.params.id).populate('ticket');

    if (!order) {
      throw new NotFoundError();
    }

    if (order.userId !== req.currentUser!.id) {
      throw new NotAuthorizedError();
    }

    res.send(order);
  }
);

export { router as showOrderRouter };
